mt_project_cluster/requirements.txt
transformers
accelerate
peft
trl
bitsandbytes
sentencepiece
datasets
sacrebleu
python-docx
pypdf
pandas
openpyxl
lxml
sentence-transformers